# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Colab Server API
  version: 0.1.4-draft
  summary: Colab 서버 REST API — P1~P3(수직 슬라이스 · 협업 코어 · HITL/재개/예산)에서 웹(W)과 colab CLI(C)가 호출하는 전부.
  description: |
    # 읽는 법

    - **SSOT**: 리소스·필드·ENUM 값은 `server/migrations/0001_init.sql`(스키마 v0)과 정확히 일치한다. 동작은 `PRD.md` FR-1~FR-9, 화면이 부르는 순서는 `SCREEN.md` §4.
    - **권한**은 각 operation `description`의 **첫 줄**에 `권한:`으로 적는다. 원칙(PRD §9): 판정은 항상 체인 최상단의 **사람 originator** 기준이며 에이전트 경유로 권한이 상승하지 않는다. 세션 내 사람 역할은 Director · deputy · 워크스페이스 멤버 셋뿐이다(FR-5.3).
    - **데몬 엔드포인트**(claim · heartbeat · event push · 토큰 폐기 통보 · 페어링 토큰 교환)는 이 문서에 없다 — `contracts/daemon-protocol.md`.
    - **colab CLI가 쓰는 operation**은 `x-colab-cli: <명령>` 확장으로 표시하고 보안 스킴 `TaskToken`(`COLAB_TASK_TOKEN`, task 범위·단기)을 쓴다. 토큰이 재큐잉으로 폐기되면 `401`이다(FR-9.1).
    - **오류**는 전부 RFC 9457 `application/problem+json`(`Problem`). 검증 실패는 `422`, 상태 전이 불가·중복·차단은 `409`, 권한 없음은 `403`(버튼은 보이되 비활성 — 사유는 `detail`).
    - **페이지네이션**은 커서 방식(`cursor` · `limit` → `next_cursor`). 메시지 타임라인만 `before`/`after` 양방향.
    - **멱등키** `Idempotency-Key`(UUID)는 메시지 게시(`POST /sessions/{id}/messages`, `POST /lanes/{id}/restart`)와 HITL 응답(`POST /hitl-requests/{id}/response`)에 **필수**, 그 밖의 생성 POST에서 선택. 같은 키의 재요청은 첫 응답을 그대로 돌려준다(`Idempotent-Replayed: true`).
    - **실시간**은 SSE 하나(`GET /workspaces/{workspaceId}/stream`). 이벤트 종류는 `StreamEvent.type`에 전부 나열한다.
    - `x-prd`: 그 operation이 구현하는 PRD 항목. `x-screen`: 부르는 화면. `x-phase`: 처음 필요한 단계(P1~P5).
  license:
    name: Proprietary
...
servers:
...
security:
  - UserSession: []
...
paths:
...
  /invites/{inviteToken}:
    parameters:
      - $ref: '#/components/parameters/InviteToken'
    get:
      tags: [auth]
      operationId: previewInvite
      summary: 초대 미리보기(S3)
      description: |
        권한: 없음(공개 — 토큰이 곧 권한).
        워크스페이스·초대자·부여될 역할을 보여준다. 만료·취소된 토큰은 `410`이고 `Problem.code`가 `invite_expired` / `invite_revoked`다.
      security: []
      x-screen: [S3]
      x-phase: P1
      responses:
        '200':
          description: 초대 정보.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InvitePreview' }
        '404': { $ref: '#/components/responses/NotFound' }
        '410': { $ref: '#/components/responses/Gone' }
        default: { $ref: '#/components/responses/Problem' }
  /invites/{inviteToken}/accept:
    parameters:
      - $ref: '#/components/parameters/InviteToken'
    post:
      tags: [auth]
      operationId: acceptInvite
      summary: 초대 수락(S3)
      description: |
        권한: 로그인한 사용자.
        초대의 워크스페이스에 `member`로 가입한다. 이미 멤버면 `200`으로 기존 멤버십을 돌려준다(멱등).
      x-prd: [P1 DoD 초대 링크]
      x-screen: [S3]
      x-phase: P1
      responses:
        '200':
          description: 이미 멤버였거나 지금 가입됨.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Member' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '410': { $ref: '#/components/responses/Gone' }
        default: { $ref: '#/components/responses/Problem' }

  # ───────────────────────────── workspaces ─────────────────────────────
  /workspaces:
    get:
      tags: [workspaces]
      operationId: listWorkspaces
      summary: 내 워크스페이스 목록
      description: |
        권한: 로그인한 사용자(본인이 멤버인 것만).
      x-phase: P1
      responses:
        '200':
          description: 목록.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Workspace' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [workspaces]
      operationId: createWorkspace
      summary: 워크스페이스 생성(S4 1단계)
      description: |
        권한: 로그인한 사용자. 생성자가 `owner`.
        `workspace_settings` 행이 PRD §7 기본값으로 함께 만들어진다.
      x-screen: [S4]
      x-phase: P1
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyOptional'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, minLength: 1, maxLength: 80 }
                slug:
                  type: string
                  pattern: '^[a-z0-9](?:[a-z0-9-]{0,38}[a-z0-9])?$'
                  description: 생략하면 이름에서 만든다.
      responses:
        '201':
          description: 생성됨.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Problem' }
  /workspaces/{workspaceId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      tags: [workspaces]
      operationId: getWorkspace
      summary: 워크스페이스 조회
      description: |
        권한: 워크스페이스 멤버.
      x-phase: P1
      responses:
        '200':
          description: 워크스페이스.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
    patch:
      tags: [workspaces]
      operationId: updateWorkspace
      summary: 워크스페이스 이름 변경
      description: |
        권한: owner · admin.
      x-phase: P5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, minLength: 1, maxLength: 80 }
      responses:
        '200':
          description: 변경됨.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workspace' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Problem' }
...
  /workspaces/{workspaceId}/members:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      tags: [workspaces]
      operationId: listMembers
      summary: 멤버 목록(S14 멤버 탭 · Director 선택 · 멘션 자동완성)
      description: |
        권한: 워크스페이스 멤버.
      x-screen: [S14, S6, S7]
      x-phase: P1
      parameters:
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: 멤버 페이지.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    properties:
                      items:
                        type: array
                        items: { $ref: '#/components/schemas/Member' }
        '403': { $ref: '#/components/responses/Forbidden' }
        default: { $ref: '#/components/responses/Problem' }
...
  /workspaces/{workspaceId}/invites:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
    get:
      tags: [workspaces]
      operationId: listInvites
      summary: 초대 목록(대기 · 만료)
      description: |
        권한: owner · admin.
      x-screen: [S14]
      x-phase: P1
      responses:
        '200':
          description: 초대 목록.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Invite' }
        '403': { $ref: '#/components/responses/Forbidden' }
        default: { $ref: '#/components/responses/Problem' }
    post:
      tags: [workspaces]
      operationId: createInvite
      summary: 초대 링크 생성
      description: |
        권한: owner · admin.
        이메일을 주면 그 주소로만 수락할 수 있고 알림 메일을 보낸다. 비우면 링크를 아는 누구나 수락한다. 기본 만료 7일.
      x-prd: [P1 DoD 초대 링크]
      x-screen: [S14, S4]
      x-phase: P1
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyOptional'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: [string, 'null'], format: email }
                role:
                  allOf: [{ $ref: '#/components/schemas/MemberRole' }]
                  default: member
                  description: owner는 초대로 줄 수 없다(`422`).
                expires_in_hours: { type: integer, minimum: 1, maximum: 720, default: 168 }
      responses:
        '201':
          description: 생성됨. `url`을 복사해 전달한다.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invite' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '422': { $ref: '#/components/responses/ValidationError' }
        default: { $ref: '#/components/responses/Problem' }
  /workspaces/{workspaceId}/invites/{inviteId}:
    parameters:
      - $ref: '#/components/parameters/WorkspaceId'
      - name: inviteId
        in: path
        required: true
        schema: { type: string, format: uuid }
    delete:
      tags: [workspaces]
      operationId: revokeInvite
      summary: 초대 취소
      description: |
        권한: owner · admin.
      x-screen: [S14]
      x-phase: P1
      responses:
        '204': { description: 취소됨. 이후 미리보기는 `410 invite_revoked`. }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        default: { $ref: '#/components/responses/Problem' }
...
components:
...
  securitySchemes:
    UserSession:
      type: apiKey
      in: cookie
      name: colab_session
      description: 로그인(S1/S2)으로 발급되는 사용자 세션 쿠키(HttpOnly · SameSite=Lax). 웹 클라이언트 기본 인증. 사람의 모든 권한 판정은 이 사용자 = originator 기준.
    TaskToken:
      type: http
      scheme: bearer
      bearerFormat: COLAB_TASK_TOKEN
      description: |
        task마다 발급되는 단기 토큰(`COLAB_TASK_TOKEN`, 형식은 `contracts/colab-cli.md`). 그 task의 세션·에이전트 범위 밖에는 접근할 수 없다(FR-7.4). 재큐잉·취소 시 서버가 폐기하며(FR-9.1) 이후 호출은 `401 token_revoked`. 테스트 채팅에는 발급되지 않는다(FR-1.8.1). `x-colab-cli`가 붙은 operation만 이 스킴을 받는다.
    DaemonToken:
      type: http
      scheme: bearer
      bearerFormat: COLAB_DAEMON_TOKEN
      description: |
        페어링(`daemon-protocol.md` §2)으로 런타임마다 발급되는 데몬 토큰(`cdt_…`). 데몬 API(`/v1/daemon/*`)는 이 문서가 아니라 `daemon-protocol.md`가 스펙이므로 여기 등장하는 operation은 **`downloadArtifact` 하나**다(v0.7.3, T-I4): §4.3 `rebind_prepare`가 데몬에게 아티팩트 다운로드를 지시하기 때문이다. 범위는 **그 런타임에 고정된 세션**의 아티팩트뿐이고, 런타임 삭제·토큰 회수로 폐기된다.

  # ───────────────────────────── parameters ─────────────────────────────
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    InviteToken:
      name: inviteToken
      in: path
      required: true
      schema: { type: string }
...
    Cursor:
      name: cursor
      in: query
      description: 이전 응답의 `next_cursor`. 불투명 문자열.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
    IdempotencyKeyOptional:
      name: Idempotency-Key
      in: header
      required: false
      description: 선택. 주면 `IdempotencyKeyRequired`와 같은 규칙.
      schema: { type: string, format: uuid }
...
  responses:
    Problem:
      description: 그 밖의 오류. 본문은 항상 `application/problem+json`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Unauthorized:
      description: "인증 없음 · 사용자 세션 만료 · TaskToken 폐기(`code: token_revoked`) 또는 만료."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            revoked:
              value: { type: 'https://colab.dev/problems/token-revoked', title: Task token revoked, status: 401, code: token_revoked, detail: 이 task는 재큐잉되어 토큰이 폐기되었다. 즉시 종료하라. }
    Forbidden:
      description: 권한 없음. 화면은 버튼을 숨기지 않고 비활성+`detail`을 보여준다(SCREEN §7). deputy 시점 제한이면 `can_respond_from`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            deputy_too_early:
              value: { type: 'https://colab.dev/problems/deputy-not-yet', title: Deputy cannot respond yet, status: 403, code: deputy_not_yet, detail: 'Director 응답 대기 중 · 14:30부터 승인 가능', can_respond_from: '2026-09-06T14:30:00Z' }
    NotFound:
      description: 없음(또는 다른 워크스페이스의 리소스 — 존재를 노출하지 않는다).
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    Conflict:
      description: 상태 전이 불가 · 유일성 충돌 · 차단(활성 세션이 걸린 런타임, 더티 workdir, 열린 HITL 존재 등). `code`로 구분.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            runtime_has_active_sessions:
              value: { type: 'https://colab.dev/problems/runtime-has-active-sessions', title: Runtime has active sessions, status: 409, code: runtime_has_active_sessions, detail: 먼저 세션을 재바인딩하거나 종료하라., sessions: [{ id: 6f1a0000-0000-4000-8000-000000000001, title: 시장 조사 }] }
    Gone:
      description: "만료·취소된 토큰/초대/페어링(`code: invite_expired` · `invite_revoked` · `pairing_expired` · `test_chat_closed`)."
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: 검증 실패. 필드별 사유는 `errors[]`.
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          examples:
            session_create:
              value: { type: 'https://colab.dev/problems/validation', title: Validation failed, status: 422, code: validation_failed, errors: [{ field: runtime_id, code: required_for_isolation, message: worktree 격리는 실행 머신이 필수다. }, { field: completion_condition, code: criteria_met_alone, message: criteria_met은 단독으로 쓸 수 없다. }] }
...
  schemas:
...
    Problem:
      type: object
      description: RFC 9457 Problem Details + `code`(기계용 식별자) + `errors[]`(필드 검증) 확장.
      required: [title, status]
      properties:
        type: { type: string, format: uri, default: 'about:blank' }
        title: { type: string }
        status: { type: integer }
        detail: { type: string, description: 사람에게 보여줄 사유(화면의 비활성 툴팁·경고 문구). }
        instance: { type: string, format: uri }
        code:
          type: string
          description: 기계용 식별자(snake_case). 예 — `token_revoked` · `deputy_not_yet` · `hitl_already_open` · `not_participant` · `runtime_offline` · `no_runtime` · `workdir_dirty` · `runtime_has_active_sessions` · `invalid_transition` · `idempotency_key_reused` · `validation_failed` · `not_designated_reviewer` · `workdir_quota_exceeded` · `running_lanes`.
        errors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: { type: string, description: JSON pointer 또는 필드명. }
              code: { type: string }
              message: { type: string }
        can_respond_from:
          type: [string, 'null']
          format: date-time
          description: "deputy 의 시점 제한(기한 절반 경과 시각)일 때만 시각. **권한이 영영 없는 사람(일반 멤버 등)에게는 비운다(null)** — 생기지 않을 권리에 시각을 약속하지 않는다(E7-11, PR #108)."
        sessions:
          type: array
          description: 차단 사유가 된 세션(런타임 삭제 409 등).
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              title: { type: string }
        workdirs:
          type: array
          description: 차단 사유가 된 작업 폴더(deleteSession 409 `workdir_unmerged`).
          items: { $ref: '#/components/schemas/Workdir' }
      additionalProperties: true
    Page:
      type: object
      description: 커서 페이지 공통 봉투. `items`는 각 operation이 정한다.
      required: [items, next_cursor]
      properties:
        items: { type: array, items: {} }
        next_cursor:
          type: [string, 'null']
          description: 다음 페이지 커서. null이면 끝.

    # ── ENUM (0001_init.sql과 동일) ──
    MemberRole:
      type: string
      enum: [owner, admin, member]
      description: '`member_role`'
...
    SessionStatus:
      type: string
      enum: [draft, active, paused, completing, completed, cancelled]
      description: '`session_status` (FR-2.3)'
...
    WorkdirKind:
      type: string
      enum: [worktree, container, dir]
      description: '`workdir_kind`'
    WorkdirStatus:
      type: string
      enum: [active, retained, deleted]
      description: '`workdir_status`'
...
    User:
      type: object
      required: [id, email, display_name, created_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        avatar_url: { type: [string, 'null'], format: uri }
        created_at: { type: string, format: date-time }
...
    Workspace:
      type: object
      required: [id, name, slug, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Member:
      type: object
      required: [id, workspace_id, user, role, created_at]
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        user: { $ref: '#/components/schemas/User' }
        role: { $ref: '#/components/schemas/MemberRole' }
        created_at: { type: string, format: date-time }
    Invite:
      type: object
      required: [id, workspace_id, role, token, url, expires_at, created_at, status]
      properties:
        id: { type: string, format: uuid }
        workspace_id: { type: string, format: uuid }
        email: { type: [string, 'null'], format: email }
        role: { $ref: '#/components/schemas/MemberRole' }
        token: { type: string }
        url: { type: string, format: uri, description: 복사해서 전달할 링크. }
        invited_by: { $ref: '#/components/schemas/User' }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
        accepted_at: { type: [string, 'null'], format: date-time }
        status:
          type: string
          enum: [pending, accepted, expired, revoked]
    InvitePreview:
      type: object
      required: [token, workspace, role, invited_by, expires_at]
      properties:
        token: { type: string }
        workspace: { $ref: '#/components/schemas/Workspace' }
        role: { $ref: '#/components/schemas/MemberRole' }
        invited_by: { $ref: '#/components/schemas/User' }
        expires_at: { type: string, format: date-time }
...
    Workdir:
      type: object
      required: [id, session_id, kind, path_or_ref, status, disk_bytes, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        session_id: { type: string, format: uuid }
        session: { $ref: '#/components/schemas/SessionRef' }
        agent_id: { type: [string, 'null'], format: uuid, description: worktree — 에이전트당 1개. }
        lane_id: { type: [string, 'null'], format: uuid, description: container · none — lane당 1개. }
        kind: { $ref: '#/components/schemas/WorkdirKind' }
        path_or_ref: { type: string }
        branch: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/WorkdirStatus' }
        disk_bytes: { type: integer, format: int64 }
        last_used_at: { type: [string, 'null'], format: date-time }
        retain_until: { type: [string, 'null'], format: date-time }
        dirty:
          type: [boolean, 'null']
          description: 미병합 커밋 또는 미커밋 변경이 있음(GC 차단 · 수동 삭제 경고, FR-6.4 M4). 데몬 마지막 보고 기준.
        merged:
          type: [boolean, 'null']
          description: "P4(T-S9 질문 2): 브랜치가 기본 브랜치에 병합됐는가. 데몬 마지막 보고 기준."
        commits_ahead:
          type: [integer, 'null']
          description: "P4: 기본 브랜치 대비 이 브랜치의 커밋 수. 0 이면 E13-11(커밋 0 + 클린 = 삭제)."
        gc_blocked_reason:
          type: [string, 'null']
          enum: [unmerged_commits, uncommitted_changes, null]
          description: "P4: 마지막 GC 판정이 삭제를 막은 사유. Director 의 다음 행동이 다르다 — `unmerged_commits` 는 병합해라(E13-12), `uncommitted_changes` 는 커밋하거나 버려라(E13-13). null = 차단 없음. deleteWorkdir 409 의 Problem.detail 에도 같은 값."
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    # ── 에이전트 ──
...
    SessionRef:
      type: object
      required: [id, title, status]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        status: { $ref: '#/components/schemas/SessionStatus' }
...